+++ SCP330Behavior.cs
@@ -38,7 +38,6 @@ namespace SCP956
                 HUDManager.Instance.UIAudio.PlayOneShot(BoneCracksfx, 1f);
                 HUDManager.Instance.DisplayTip("Took too much candy", "You feel a sharp pain where your hands should be. They've been severed by an unknown force.");
                 // TODO: Make it so the players hands are no longer visible
-                // TODO: Make player drunk?
 
                 StatusEffectController.Instance.DamagePlayerOverTime(5, 2, true);