+++ StatusEffectController.cs
@@ -321,7 +321,7 @@ namespace SCP956
         {
             HUDManager.Instance.UIAudio.PlayOneShot(Plugin.CandleBlowsfx, 1f);
             yield return new WaitForSecondsRealtime(1f);
-            Plugin.PlayerAge = 10;
+            Plugin.PlayerAge = 11;
             NetworkHandler.Instance.ChangePlayerSizeServerRpc(localPlayer.actualClientId, 0.8f);
             HUDManager.Instance.UIAudio.PlayOneShot(Plugin.CakeAppearsfx, 1f);
         }